Make PredicateGenerator.parsePredicateValue package protected
* Move a unit test from AdminProgramBlockPredicatesControllerTest to PredicateGeneratorTest because it's testing PredicateGenerator
* Reduce visibility of PredicateGenerator.parsePredicateValue to package protected because it's now just used privately and in the unit tests
